refactor & document GameEngine, AB#427
