+++ .idea/compiler.xml
@@ -7,6 +7,7 @@
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
         <module name="The-Lone-Survivor" />
+        <module name="Lone-Survivor" />
       </profile>
     </annotationProcessing>
   </component>
