+++ src/main/resources/json/Outside.json
@@ -1,0 +1,44 @@
+[
+  {
+    "locationName": "outside plane",
+    "locationDescription": "You are outside the plane crash. Dense tropical rainforest surround the area.",
+    "locationItems": [],
+    "locationDirections": {"north": "plane", "south": "forest", "east":  "???", "west":  "???"}
+  },
+  {
+    "locationName": "forest",
+    "locationDescription": "Dense tropical rainforest surround the area. You hear sounds of water nearby.",
+    "locationItems": [],
+    "locationDirections": {"north": "outside plane", "south": "bridge", "east":  "???", "west":  "???"}
+  },
+  {
+    "locationName": "steel bridge",
+    "locationDescription": "An old steel bridge wrapped in vegetation appears. Seems safe to cross.",
+    "locationItems": [],
+    "locationDirections": {"north": "forest", "east": "cross bridge"}
+  },
+  {
+    "locationName": "East of bridge",
+    "locationDescription": "You make it across the bridge safely. You hear sounds of unintelligible human voices nearby.",
+    "locationItems": [],
+    "locationDirections": {"west": "cross bridge", "south": "investigate noises"}
+  },
+  {
+    "locationName": "local village",
+    "locationDescription": "You come across a local village and see human faces about. You are relieved. ",
+    "locationItems": [],
+    "locationDirections": {"north": "steel bridge"}
+  },
+  {
+    "locationName": "mysterious animal",
+    "locationDescription": "You hear low growling noises, and you see what looks to be a mutated bear and probably the result of a bad science experiment! ",
+    "locationItems": [],
+    "locationDirections": {"north": "forest"}
+  },
+  {
+    "locationName": "forest dead end",
+    "locationDescription": "The forest seems too dense to go any further at this point. Might be best to re-trace your steps.",
+    "locationItems": [],
+    "locationDirections": {"north": "economy class"}
+  }
+]
