+++ src/main/resources/json/inventory.json
